+++ Thrust_lib/unsigned_distance_function.h
@@ -13,8 +13,8 @@ double unsigned_distance_space_map_cuda(const Points& points, const Point& targe
 double unsigned_distance_brute_force(const Faces& faces, 
 	const Points& points, const Point& target, double beta, int& nearest_point);
 
-void print_output(float dist, int nearest_point, const Point& target, const Points& points);
-void print_output(float dist, int nearest_point, const Point& target, const Points& points, const Faces& faces);
+void print_output(double dist, int nearest_point, const Point& target, const Points& points);
+void print_output(double dist, int nearest_point, const Point& target, const Points& points, const Faces& faces);
 void test_local();
 
 // Run program: Ctrl + F5 or Debug > Start Without Debugging menu
